Update horizontal-vertical-gallery.md
This PR improves the responsiveness of the "Horizontal & Vertical Gallery" CSS snippet. On smaller screens (e.g., mobile devices), the layout adjusts from a grid view to a stacked and scrollable thumbnail view. This ensures better usability and accessibility on all devices. Thank you for maintaining such an excellent resource!

Signed-off-by: Stuti Porwal <porwalstuti01@gmail.com>